refactor: change file name and update image test
